11/1-12pm

+++ WebContent/notice/noticeSelectResult.jsp
@@ -5,7 +5,7 @@
 	request.setCharacterEncoding("UTF-8");
 	response.setCharacterEncoding("UTF-8");
 	
-	NoticeDTO dto = (NoticeDTO)request.getAttribute("DTO");
+	NoticeDTO noticedto = (NoticeDTO)request.getAttribute("DTO");
 
 
 
@@ -13,7 +13,7 @@
 <!DOCTYPE html>
 <html>
 <head>
-<title>Bootstrap Example</title>
+<title>S-R-page</title>
 <meta charset="utf-8">
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <link rel="stylesheet"
@@ -35,6 +35,7 @@
 
 h2 {
 	text-align: center;
+	margin-bottom: 25px;
 }
 #contents{
 	min-height: 500px;
@@ -42,20 +43,21 @@ h2 {
 }
 
 #bt_gp{
-	margin-left: 160px;
+	margin-left: 17%;
+	
 }
 </style>
 </head>
 <body>
-
+<%@ include file="../layout/nav.jsp" %>
 	<div class="container">
-		<h2>NOTICE</h2>
+		<h2>NOTICE S-R-page</h2>
 		<form class="form-horizontal" action="/action_page.php">
 			<div class="form-group">
 				<label class="control-label col-sm-2" for="title">TITLE</label>
 				<div class="col-sm-10">
 					<input type="text" class="form-control" id="title" name="title"
-						readonly="readonly" value="<%=dto.getTitle()%>">
+						readonly="readonly" value="<%=noticedto.getTitle()%>">
 				</div>
 			</div>
 
@@ -63,21 +65,21 @@ h2 {
 				<label class="control-label col-sm-2" for="title" id="hit_lb">HIT</label>
 				<div class="col-sm-10">
 					<input type="text" class="form-control" id="hit" name="hit"
-						readonly="readonly" value="<%=dto.getHit() %>">
+						readonly="readonly" value="<%=noticedto.getHit() %>">
 				</div>
 			</div>
 			<div class="form-group">
 				<label class="control-label col-sm-2" for="title">WRITER</label>
 				<div class="col-sm-10">
 					<input type="text" class="form-control" id="writer" name="writer"
-						readonly="readonly" value="<%=dto.getWriter() %>">
+						readonly="readonly" value="<%=noticedto.getWriter() %>">
 				</div>
 			</div>
 			<div class="form-group">
 				<label class="control-label col-sm-2" for="title">CONTENTS</label>
 				<div class="col-sm-10">
 					<textarea class="form-control" id="contents" name="contents"
-						readonly="readonly"><%=dto.getContents() %></textarea>
+						readonly="readonly"><%=noticedto.getContents() %></textarea>
 				</div>
 			</div>
 		</form>
@@ -85,15 +87,18 @@ h2 {
 		<!--**********************버튼그룹 -->
 
 <div id = "bt_gp">
+		<%if(dto !=null && dto.getLev()==0) {%>
 		<button type="button" class="btn btn-info">글쓰기</button>
+		<%if(dto.getId().equals(noticedto.getWriter())) {%>
 		<button type="button" class="btn btn-warning">수정</button>
 		<button type="button" class="btn btn-danger">삭제</button>
+		<%}} %>
 		<button type="button" class="btn btn-primary" id="btn_right">목록</button>
 </div>
 	</div>
 	<script type="text/javascript">
 	
-	var num = <%= dto.getNum() %>;
+	var num = <%= noticedto.getNum() %>;
 	
 	$(".btn-warning").click(function() {
 		location.href="noticeUpdate.jsp?num="+num;
@@ -110,8 +115,12 @@ h2 {
 	});
 	
 	$(".btn-danger").click(function() {
+		if (confirm("삭제하시겠습니까?")) {
 		location.href="noticeDelete.jsp?num="+num;
+		
+		}
 	});
+	
 </script>
 </body>
 </html>
